Redesign app and remove the MainActivity
