+++ app/src/main/AndroidManifest.xml
@@ -9,20 +9,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.graphhopper.navigation.example.MainActivity">
+
+        <activity android:name="com.graphhopper.navigation.example.NavigationLauncherActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-
-        <activity
-            android:name="com.graphhopper.navigation.example.NavigationLauncherActivity"
-            android:label="">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.graphhopper.navigation.example.MainActivity" />
         </activity>
 
         <activity
