--- app/src/main/java/com/graphhopper/navigation/example/SampleItem.java
@@ -1,27 +1,0 @@
-package com.graphhopper.navigation.example;
-
-public class SampleItem {
-
-    private String name;
-    private String description;
-    private Class activity;
-
-    public SampleItem(String name, String description, Class activity) {
-        this.name = name;
-        this.description = description;
-        this.activity = activity;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public Class getActivity() {
-        return activity;
-    }
-
-}
