+++ app/src/main/res/drawable/ic_navigation_black_24dp.xml
@@ -1,7 +1,7 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="24dp"
     android:height="24dp"
-    android:tint="#D1D1D1"
+    android:tint="#00f148"
     android:viewportHeight="24.0"
     android:viewportWidth="24.0">
     <path
