+++ app/src/main/res/values/strings.xml
@@ -1,9 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <string name="app_name" tools:override="true">GraphHopper Demo</string>
 
-    <string name="title_navigation_launcher">Navigation Sample</string>
-    <string name="description_navigation_launcher">Shows basic navigation with multiple waypoints and fetches solutions from the GraphHopper Route Optimization API</string>
-
     <string name="settings">Settings</string>
     <string name="simulate_route">Simulate Route</string>
     <string name="language">Language</string>
